// src/lib/types.ts
export interface TranscriptSegment {
  text: string;
  start: number;
  duration: number;
}

export interface CaptionTrack {
  languageCode: string;
  languageName: string;
  isAutoGenerated: boolean;
}

export interface TranscriptResponse {
  videoId: string;
  title: string;
  segments: TranscriptSegment[];
  availableTracks: CaptionTrack[];
  selectedTrack: CaptionTrack;
}